Adds google analytics, and moves tests to their own directories
